backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
bcrypt
python-jose[cryptography]
python-dotenv
email-validator
resend
